media-source/SourceBuffer-appendStream_removed.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SourceBuffer#appendStream(stream) after removing the sourceBuffer object</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'stream';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(xhr.response);
    };
}
//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    GET(FILE, function(stream) {
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e){
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            mediaSource.removeSourceBuffer(sourceBuffer);
            t.step(function() {
                assert_throws({name: 'InvalidStateError'},
                              function() {
                                  sourceBuffer.appendStream(stream);
                              });
            });
            t.done();
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
},
'SourceBuffer#appendStream(stream) : If the SourceBuffer has been removed from ' +
'the sourceBuffers attribute of the parent media source, INVALID_STATE_ERR exception should be thrown.');

</script>
</body>
</html>
